design #7: css foundation 수정 및 프로젝트 공통 css 수정
