+++ fe/src/styles/base/font.ts
@@ -1,4 +1,4 @@
-const FONT_BASE = {
+export const FONT_BASE = {
   bold32: '700 32px/48px Pretendard Variable, sans-serif',
   bold24: '700 24px/36px Pretendard Variable, sans-serif',
   bold20: '700 20px/32px Pretendard Variable, sans-serif',
@@ -9,20 +9,4 @@ const FONT_BASE = {
   medium20: '500 20px/32px Pretendard Variable, sans-serif',
   medium16: '500 16px/24px Pretendard Variable, sans-serif',
   medium12: '500 12px/16px Pretendard Variable, sans-serif',
-};
-
-export const font = {
-  displayBold32: FONT_BASE.bold32,
-  displayBold20: FONT_BASE.bold20,
-  displayBold16: FONT_BASE.bold16,
-  displayBold12: FONT_BASE.bold12,
-  displayMedium20: FONT_BASE.medium20,
-  displayMedium16: FONT_BASE.medium16,
-  displayMedium12: FONT_BASE.medium12,
-  selectedBold20: FONT_BASE.bold20,
-  selectedBold16: FONT_BASE.bold16,
-  selectedBold12: FONT_BASE.bold12,
-  availableMedium20: FONT_BASE.medium20,
-  availableMedium16: FONT_BASE.medium16,
-  availableMedium12: FONT_BASE.medium12,
 };
