+++ fe/src/styles/index.css
@@ -3,11 +3,9 @@
 body {
   font-family: 'Pretendard Variable', Pretendard, Inter, system-ui, Avenir,
     Helvetica, Arial, sans-serif;
-  background-color: #f7f7fc;
 
   & * {
     padding: 0;
-    box-sizing: border-box;
   }
 
   & button {
