--- fe/src/styles/object.ts
@@ -1,23 +1,0 @@
-export const radius = {
-  half: '50%',
-  medium: '12px',
-  large: '16px',
-};
-
-export const border = {
-  default: '1px solid',
-  icon: '1.6px solid',
-  dash: '1px dash',
-};
-
-export const opacity = {
-  default: '1',
-  hover: '0.8',
-  press: '0.64',
-  disabled: '0.32',
-};
-
-export const shadow = {
-  light: '0px 0px 8px 0px #14142B0A',
-  dark: '0px 0px 16px 0px #14142BCC',
-};
